latticearc::primitives::fips_error

Trait FipsError

pub trait FipsError {
    // Required method
    fn fips_code(&self) -> FipsErrorCode;

    // Provided methods
    fn fips_message(&self) -> String { ... }
    fn is_fips_critical(&self) -> bool { ... }
}

Trait for errors that can be converted to FIPS error codes.

Implement this trait for error types to provide FIPS-compliant error reporting. The implementation should map each error variant to the most appropriate FIPS error code.

Example

use latticearc::primitives::fips_error::{FipsError, FipsErrorCode};

#[derive(Debug)]
enum MyError {
    BadKey,
    Timeout,
}

impl FipsError for MyError {
    fn fips_code(&self) -> FipsErrorCode {
        match self {
            MyError::BadKey => FipsErrorCode::InvalidKeyLength,
            MyError::Timeout => FipsErrorCode::Timeout,
        }
    }
}

Required Methods

fn fips_code(&self) -> FipsErrorCode

Returns the FIPS error code for this error.

Provided Methods

fn fips_message(&self) -> String

Returns the FIPS-formatted error string.

This is a convenience method that formats the error code for logging.

fn is_fips_critical(&self) -> bool

Checks if this error is critical and requires module shutdown.