latticearc::unified_api::audit

Trait AuditStorage

pub trait AuditStorage: Send + Sync {
    // Required methods
    fn write(&self, event: &AuditEvent) -> Result<()>;
    fn flush(&self) -> Result<()>;
}

Trait for audit storage implementations.

Implement this trait to create custom audit storage backends (e.g., database, remote service, etc.).

Required Methods

fn write(&self, event: &AuditEvent) -> Result<()>

Write an audit event to storage.

Errors

Returns an error if the event cannot be written.

fn flush(&self) -> Result<()>

Flush any buffered events to persistent storage.

Errors

Returns an error if the flush operation fails.